# server-side/swagger.yaml
openapi: 3.0.0
info:
  title: shop24 v2.0
  version: 1.0.0
  description: ''
servers:
  - url: '{{baseURL}}'
  - url: http://localhost:5000
paths:
  /auth/register:
    post:
      summary: register
      tags:
        - auth
      responses: {}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
  /auth/login:
    post:
      summary: login
      tags:
        - auth
      responses: {}
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer